🌵 Added the ability to configure add-ons
